pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "usdc-weth-model"
version = "0.1.0"
description = ''
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
  { name = "Dave Bryson", email = "davebryson@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
  "matplotlib>=3.8.2",
  "simular-evm>=0.2.2",
  "mesa>=2.2.4",
  "pytest>=8.1.1",
  "jupyter>=1.0.0",
  "scipy>=1.13.0",
  "tqdm>=4.66.2",
]
